+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - **scoop-download**: Fix function `nightly_version` not defined error ([#6386](https://github.com/ScoopInstaller/Scoop/issues/6386))
 
+### Code Refactoring
+
+- **output**: replace raw prints with functions for standardized output
+
 ## [v0.5.3](https://github.com/ScoopInstaller/Scoop/compare/v0.5.2...v0.5.3) - 2025-08-11
 
 ### Features