+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Code Refactoring
 
-- **output**: replace raw prints with functions for standardized output
+- **output**: replace raw prints with functions for standardized output ([#6450](https://github.com/ScoopInstaller/Scoop/issues/6450))
 
 ## [v0.5.3](https://github.com/ScoopInstaller/Scoop/compare/v0.5.2...v0.5.3) - 2025-08-11