+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Features
 
-**autoupdate:** GitHub predefined hashes support ([#6416](https://github.com/ScoopInstaller/Scoop/issues/6416), [#6435](https://github.com/ScoopInstaller/Scoop/issues/6435))
+- **autoupdate:** GitHub predefined hashes support ([#6416](https://github.com/ScoopInstaller/Scoop/issues/6416), [#6435](https://github.com/ScoopInstaller/Scoop/issues/6435))
+- **install:** Add output for the setting and removal of environment variables ([#6459](https://github.com/ScoopInstaller/Scoop/issues/6459))
 
 ### Bug Fixes