+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 - **autoupdate:** Use origin URL to handle URLs with fragment in GitHub mode ([#6455](https://github.com/ScoopInstaller/Scoop/issues/6455))
 - **buckets|scoop-info:** Switch git log date format to ISO 8601 to avoid locale issues ([#6446](https://github.com/ScoopInstaller/Scoop/pull/6446))
 
+### Code Refactoring
+
+- **output**: Replace raw prints with functions for standardized output ([#6450](https://github.com/ScoopInstaller/Scoop/issues/6450))
+
 ## [v0.5.3](https://github.com/ScoopInstaller/Scoop/compare/v0.5.2...v0.5.3) - 2025-08-11
 
 ### Features