+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - **buckets|scoop-info:** Switch git log date format to ISO 8601 to avoid locale issues ([#6446](https://github.com/ScoopInstaller/Scoop/issues/6446))
 - **scoop-version:** Fix logic error caused by missing brackets ([#6463](https://github.com/ScoopInstaller/Scoop/issues/6463))
 
+### Code Refactoring
+
+- **output**: Replace raw prints with functions for standardized output ([#6450](https://github.com/ScoopInstaller/Scoop/issues/6450))
+
 ## [v0.5.3](https://github.com/ScoopInstaller/Scoop/compare/v0.5.2...v0.5.3) - 2025-08-11
 
 ### Features