+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased](https://github.com/ScoopInstaller/Scoop/compare/v0.5.3...develop)
 
+### Features
+
+- **download**: Block apps without file hash by default for security reasons ([#6387](https://github.com/ScoopInstaller/Scoop/issues/6387))
+
 ### Bug Fixes
 
 - **scoop-download:** Fix function `nightly_version` not defined error ([#6386](https://github.com/ScoopInstaller/Scoop/issues/6386))